refactor(metrics): extract internal/metrics package from state
Carve performance-metric and retrospective persistence out of internal/state
into a dedicated internal/metrics sibling package. State stays focused on
run/step lifecycle; metrics owns the performance_metric and retrospective
tables. Schema migrations remain in state's migration runner since there is
one runner per database.

What moved:
- PerformanceMetricRecord, PerformanceQueryOptions, StepPerformanceStats,
  RetrospectiveRecord, ListRetrosOptions types (state/types.go,store.go ->
  metrics/types.go).
- RecordPerformanceMetric, GetPerformanceMetrics, GetStepPerformanceStats,
  GetRecentPerformanceHistory, CleanupOldPerformanceMetrics SQL methods
  (state/metrics.go -> metrics/store.go).
- SaveRetrospective, GetRetrospective, ListRetrospectives, DeleteRetrospective,
  UpdateRetrospectiveSmoothness, UpdateRetrospectiveStatus SQL methods
  (state/retros.go -> metrics/store.go).
- Existing round-trip tests rewritten as metrics/store_test.go on a fresh
  in-memory schema so the package tests itself with zero state coupling.

Interface changes:
- state.RunStore drops the metrics + retrospective method blocks. Consumers
  that needed them now depend on *metrics.Store directly, in line with the
  StateStore narrowing trajectory from #1496.
- New helper state.UnderlyingDB(StateStore) exposes the *sql.DB so callers
  can build a metrics.Store from the same connection pool without an
  interface explosion. metrics.NewStore takes a small DB interface that
  *sql.DB satisfies implicitly.

Caller updates:
- internal/pipeline executor stores a *metrics.Store alongside StateStore;
  WithStateStore auto-derives one from the underlying DB, plus a
  WithMetricsStore option for tests. Dispatch + contract paths now write
  via e.metrics. ETACalculator takes a small statsLookup interface so unit
  tests can stub it without a real DB.
- internal/retro Collector and Storage interfaces now reference metrics
  types. Generator constructor takes a *metrics.Store alongside the existing
  state surface.
- internal/webui retro and analytics handlers build a metrics.Store from
  the runtime store for each call.
- internal/tui persona and detail providers take a *metrics.Store; app.go
  wires it from deps.Store via UnderlyingDB.
- cmd/wave retro and run_stages helpers construct the metrics handle next
  to the existing state store.
- internal/testutil MockStateStore drops the moved methods (no callers).

Lint:
- .golangci.yml depguard rule for ADR-007 (database/sql limited to
  internal/state) extended to whitelist internal/metrics, the new sibling
  persistence package.

Tests: go build ./...; go test -race ./... all green. New
metrics/store_test.go covers record + read-back, list-by-run, retro save/get
round-trip, smoothness/status updates, and cleanup pruning.
