+++ .golangci.yml
@@ -60,17 +60,19 @@ linters:
         # These are string-literal rules depguard cannot express. See the
         # package doc in internal/ontology/service.go.
 
-        # Rule: Only internal/state may import database/sql (ADR-007).
-        # depguard rules apply to files matching `files:`; use negative globs
-        # so the rule fires everywhere *except* inside internal/state/**.
+        # Rule: Only internal/state and internal/metrics may import
+        # database/sql (ADR-007). internal/metrics was carved out of state in
+        # issue #62 and continues to own its sibling persistence layer; both
+        # speak SQLite directly while everything else stays interface-bound.
         no-direct-sql:
           list-mode: lax
           files:
             - "$all"
             - "!**/internal/state/**"
+            - "!**/internal/metrics/**"
           deny:
             - pkg: "database/sql"
-              desc: "ADR-007: only internal/state may import database/sql"
+              desc: "ADR-007: only internal/state and internal/metrics may import database/sql"
 
         # Rule 2: Infrastructure packages must not import Domain packages.
         infrastructure-no-domain:
