+++ internal/state/store_test.go
@@ -2150,352 +2150,6 @@ func TestRecordStepAttempt_NilCompletedAt(t *testing.T) {
 	assert.Nil(t, attempts[0].CompletedAt)
 }
 
-// =============================================================================
-// Performance Metrics Tests
-// =============================================================================
-
-// TestRecordPerformanceMetric tests recording and retrieving performance metrics.
-func TestRecordPerformanceMetric(t *testing.T) {
-	t.Run("round-trip record and retrieve", func(t *testing.T) {
-		store, cleanup := setupTestStore(t)
-		defer cleanup()
-
-		runID, err := store.CreateRun("test-pipeline", "input")
-		require.NoError(t, err)
-
-		startedAt := time.Now().Truncate(time.Second)
-		completedAt := startedAt.Add(5 * time.Second)
-
-		metric := &PerformanceMetricRecord{
-			RunID:              runID,
-			StepID:             "step-1",
-			PipelineName:       "test-pipeline",
-			Persona:            "craftsman",
-			StartedAt:          startedAt,
-			CompletedAt:        &completedAt,
-			DurationMs:         5000,
-			TokensUsed:         1200,
-			FilesModified:      3,
-			ArtifactsGenerated: 2,
-			MemoryBytes:        1024000,
-			Success:            true,
-			ErrorMessage:       "",
-		}
-
-		err = store.RecordPerformanceMetric(metric)
-		require.NoError(t, err)
-		assert.NotZero(t, metric.ID, "ID should be set after insert")
-
-		metrics, err := store.GetPerformanceMetrics(runID, "")
-		require.NoError(t, err)
-		require.Len(t, metrics, 1)
-
-		got := metrics[0]
-		assert.Equal(t, runID, got.RunID)
-		assert.Equal(t, "step-1", got.StepID)
-		assert.Equal(t, "test-pipeline", got.PipelineName)
-		assert.Equal(t, "craftsman", got.Persona)
-		assert.Equal(t, startedAt.Unix(), got.StartedAt.Unix())
-		require.NotNil(t, got.CompletedAt)
-		assert.Equal(t, completedAt.Unix(), got.CompletedAt.Unix())
-		assert.Equal(t, int64(5000), got.DurationMs)
-		assert.Equal(t, 1200, got.TokensUsed)
-		assert.Equal(t, 3, got.FilesModified)
-		assert.Equal(t, 2, got.ArtifactsGenerated)
-		assert.Equal(t, int64(1024000), got.MemoryBytes)
-		assert.True(t, got.Success)
-		assert.Empty(t, got.ErrorMessage)
-	})
-
-	t.Run("filter by stepID", func(t *testing.T) {
-		store, cleanup := setupTestStore(t)
-		defer cleanup()
-
-		runID, err := store.CreateRun("test-pipeline", "input")
-		require.NoError(t, err)
-
-		now := time.Now().Truncate(time.Second)
-
-		for _, stepID := range []string{"step-a", "step-a", "step-b"} {
-			err = store.RecordPerformanceMetric(&PerformanceMetricRecord{
-				RunID:        runID,
-				StepID:       stepID,
-				PipelineName: "test-pipeline",
-				StartedAt:    now,
-				Success:      true,
-			})
-			require.NoError(t, err)
-		}
-
-		// All metrics for the run
-		all, err := store.GetPerformanceMetrics(runID, "")
-		require.NoError(t, err)
-		assert.Len(t, all, 3)
-
-		// Only step-a
-		stepA, err := store.GetPerformanceMetrics(runID, "step-a")
-		require.NoError(t, err)
-		assert.Len(t, stepA, 2)
-		for _, m := range stepA {
-			assert.Equal(t, "step-a", m.StepID)
-		}
-
-		// Only step-b
-		stepB, err := store.GetPerformanceMetrics(runID, "step-b")
-		require.NoError(t, err)
-		assert.Len(t, stepB, 1)
-		assert.Equal(t, "step-b", stepB[0].StepID)
-	})
-
-	t.Run("nil CompletedAt handling", func(t *testing.T) {
-		store, cleanup := setupTestStore(t)
-		defer cleanup()
-
-		runID, err := store.CreateRun("test-pipeline", "input")
-		require.NoError(t, err)
-
-		now := time.Now().Truncate(time.Second)
-
-		err = store.RecordPerformanceMetric(&PerformanceMetricRecord{
-			RunID:        runID,
-			StepID:       "step-1",
-			PipelineName: "test-pipeline",
-			StartedAt:    now,
-			CompletedAt:  nil,
-			Success:      false,
-			ErrorMessage: "timed out",
-		})
-		require.NoError(t, err)
-
-		metrics, err := store.GetPerformanceMetrics(runID, "step-1")
-		require.NoError(t, err)
-		require.Len(t, metrics, 1)
-		assert.Nil(t, metrics[0].CompletedAt)
-		assert.False(t, metrics[0].Success)
-		assert.Equal(t, "timed out", metrics[0].ErrorMessage)
-	})
-}
-
-// TestGetStepPerformanceStats tests aggregated performance statistics.
-func TestGetStepPerformanceStats(t *testing.T) {
-	t.Run("aggregation across multiple metrics", func(t *testing.T) {
-		store, cleanup := setupTestStore(t)
-		defer cleanup()
-
-		now := time.Now().Truncate(time.Second)
-
-		// Record 3 metrics: 2 successes, 1 failure
-		durations := []int64{1000, 2000, 3000}
-		tokens := []int{100, 200, 300}
-		successes := []bool{true, true, false}
-
-		for i := 0; i < 3; i++ {
-			runID, err := store.CreateRun("my-pipeline", "input")
-			require.NoError(t, err)
-
-			err = store.RecordPerformanceMetric(&PerformanceMetricRecord{
-				RunID:              runID,
-				StepID:             "build",
-				PipelineName:       "my-pipeline",
-				Persona:            "craftsman",
-				StartedAt:          now.Add(time.Duration(i) * time.Second),
-				DurationMs:         durations[i],
-				TokensUsed:         tokens[i],
-				FilesModified:      i + 1,
-				ArtifactsGenerated: 1,
-				Success:            successes[i],
-			})
-			require.NoError(t, err)
-		}
-
-		stats, err := store.GetStepPerformanceStats("my-pipeline", "build", now.Add(-1*time.Hour))
-		require.NoError(t, err)
-		require.NotNil(t, stats)
-
-		assert.Equal(t, "build", stats.StepID)
-		assert.Equal(t, "craftsman", stats.Persona)
-		assert.Equal(t, 3, stats.TotalRuns)
-		assert.Equal(t, 2, stats.SuccessfulRuns)
-		assert.Equal(t, 1, stats.FailedRuns)
-		assert.Equal(t, int64(2000), stats.AvgDurationMs) // (1000+2000+3000)/3 = 2000
-		assert.Equal(t, int64(1000), stats.MinDurationMs)
-		assert.Equal(t, int64(3000), stats.MaxDurationMs)
-		assert.Equal(t, 200, stats.AvgTokensUsed)   // (100+200+300)/3 = 200
-		assert.Equal(t, 600, stats.TotalTokensUsed) // 100+200+300
-		assert.True(t, stats.TokenBurnRate > 0)
-	})
-
-	t.Run("empty result for non-existent step", func(t *testing.T) {
-		store, cleanup := setupTestStore(t)
-		defer cleanup()
-
-		stats, err := store.GetStepPerformanceStats("no-pipeline", "no-step", time.Now().Add(-1*time.Hour))
-		require.NoError(t, err)
-		require.NotNil(t, stats)
-		assert.Equal(t, 0, stats.TotalRuns)
-		assert.Equal(t, "no-step", stats.StepID)
-	})
-
-	t.Run("since filter excludes old metrics", func(t *testing.T) {
-		store, cleanup := setupTestStore(t)
-		defer cleanup()
-
-		oldTime := time.Now().Add(-2 * time.Hour).Truncate(time.Second)
-		recentTime := time.Now().Truncate(time.Second)
-		cutoff := time.Now().Add(-1 * time.Hour)
-
-		// Old metric
-		runID1, err := store.CreateRun("pipe", "input")
-		require.NoError(t, err)
-		err = store.RecordPerformanceMetric(&PerformanceMetricRecord{
-			RunID:        runID1,
-			StepID:       "step-x",
-			PipelineName: "pipe",
-			StartedAt:    oldTime,
-			DurationMs:   9999,
-			TokensUsed:   9999,
-			Success:      true,
-		})
-		require.NoError(t, err)
-
-		// Recent metric
-		runID2, err := store.CreateRun("pipe", "input")
-		require.NoError(t, err)
-		err = store.RecordPerformanceMetric(&PerformanceMetricRecord{
-			RunID:        runID2,
-			StepID:       "step-x",
-			PipelineName: "pipe",
-			StartedAt:    recentTime,
-			DurationMs:   500,
-			TokensUsed:   100,
-			Success:      true,
-		})
-		require.NoError(t, err)
-
-		stats, err := store.GetStepPerformanceStats("pipe", "step-x", cutoff)
-		require.NoError(t, err)
-		require.NotNil(t, stats)
-		assert.Equal(t, 1, stats.TotalRuns)
-		assert.Equal(t, int64(500), stats.AvgDurationMs)
-		assert.Equal(t, 100, stats.AvgTokensUsed)
-	})
-}
-
-// TestGetRecentPerformanceHistory tests filtered performance history retrieval.
-func TestGetRecentPerformanceHistory(t *testing.T) {
-	t.Run("limit enforcement", func(t *testing.T) {
-		store, cleanup := setupTestStore(t)
-		defer cleanup()
-
-		now := time.Now().Truncate(time.Second)
-
-		for i := 0; i < 5; i++ {
-			runID, err := store.CreateRun("pipe", "input")
-			require.NoError(t, err)
-			err = store.RecordPerformanceMetric(&PerformanceMetricRecord{
-				RunID:        runID,
-				StepID:       "step-1",
-				PipelineName: "pipe",
-				StartedAt:    now.Add(time.Duration(i) * time.Second),
-				Success:      true,
-			})
-			require.NoError(t, err)
-		}
-
-		metrics, err := store.GetRecentPerformanceHistory(PerformanceQueryOptions{
-			Limit: 2,
-		})
-		require.NoError(t, err)
-		assert.Len(t, metrics, 2)
-	})
-
-	t.Run("filter by pipeline name", func(t *testing.T) {
-		store, cleanup := setupTestStore(t)
-		defer cleanup()
-
-		now := time.Now().Truncate(time.Second)
-
-		for _, pname := range []string{"alpha", "alpha", "beta"} {
-			runID, err := store.CreateRun(pname, "input")
-			require.NoError(t, err)
-			err = store.RecordPerformanceMetric(&PerformanceMetricRecord{
-				RunID:        runID,
-				StepID:       "s1",
-				PipelineName: pname,
-				StartedAt:    now,
-				Success:      true,
-			})
-			require.NoError(t, err)
-		}
-
-		metrics, err := store.GetRecentPerformanceHistory(PerformanceQueryOptions{
-			PipelineName: "alpha",
-		})
-		require.NoError(t, err)
-		assert.Len(t, metrics, 2)
-		for _, m := range metrics {
-			assert.Equal(t, "alpha", m.PipelineName)
-		}
-	})
-
-	t.Run("filter by step ID", func(t *testing.T) {
-		store, cleanup := setupTestStore(t)
-		defer cleanup()
-
-		now := time.Now().Truncate(time.Second)
-
-		runID, err := store.CreateRun("pipe", "input")
-		require.NoError(t, err)
-
-		for _, sid := range []string{"build", "build", "test"} {
-			err = store.RecordPerformanceMetric(&PerformanceMetricRecord{
-				RunID:        runID,
-				StepID:       sid,
-				PipelineName: "pipe",
-				StartedAt:    now,
-				Success:      true,
-			})
-			require.NoError(t, err)
-		}
-
-		metrics, err := store.GetRecentPerformanceHistory(PerformanceQueryOptions{
-			StepID: "build",
-		})
-		require.NoError(t, err)
-		assert.Len(t, metrics, 2)
-	})
-
-	t.Run("filter by persona", func(t *testing.T) {
-		store, cleanup := setupTestStore(t)
-		defer cleanup()
-
-		now := time.Now().Truncate(time.Second)
-
-		runID, err := store.CreateRun("pipe", "input")
-		require.NoError(t, err)
-
-		for _, persona := range []string{"navigator", "craftsman", "navigator"} {
-			err = store.RecordPerformanceMetric(&PerformanceMetricRecord{
-				RunID:        runID,
-				StepID:       "s1",
-				PipelineName: "pipe",
-				Persona:      persona,
-				StartedAt:    now,
-				Success:      true,
-			})
-			require.NoError(t, err)
-		}
-
-		metrics, err := store.GetRecentPerformanceHistory(PerformanceQueryOptions{
-			Persona: "navigator",
-		})
-		require.NoError(t, err)
-		assert.Len(t, metrics, 2)
-		for _, m := range metrics {
-			assert.Equal(t, "navigator", m.Persona)
-		}
-	})
-}
 
 // =============================================================================
 // Progress Snapshot Tests
