+++ internal/state/types.go
@@ -138,51 +138,6 @@ type CancellationRecord struct {
 	Force       bool
 }
 
-// PerformanceMetricRecord holds historical performance data for a step.
-type PerformanceMetricRecord struct {
-	ID                 int64
-	RunID              string
-	StepID             string
-	PipelineName       string
-	Persona            string
-	StartedAt          time.Time
-	CompletedAt        *time.Time
-	DurationMs         int64
-	TokensUsed         int
-	FilesModified      int
-	ArtifactsGenerated int
-	MemoryBytes        int64
-	Success            bool
-	ErrorMessage       string
-}
-
-// PerformanceQueryOptions specifies filters for performance queries.
-type PerformanceQueryOptions struct {
-	PipelineName string
-	StepID       string
-	Persona      string
-	Since        time.Time
-	Limit        int
-}
-
-// StepPerformanceStats holds aggregated performance statistics for a step.
-type StepPerformanceStats struct {
-	StepID           string
-	Persona          string
-	TotalRuns        int
-	SuccessfulRuns   int
-	FailedRuns       int
-	AvgDurationMs    int64
-	MinDurationMs    int64
-	MaxDurationMs    int64
-	AvgTokensUsed    int
-	TotalTokensUsed  int
-	AvgFilesModified int
-	AvgArtifacts     int
-	LastRunAt        time.Time
-	TokenBurnRate    float64 // tokens per second
-}
-
 // ProgressSnapshotRecord holds a point-in-time progress snapshot.
 type ProgressSnapshotRecord struct {
 	ID              int64
@@ -282,17 +237,6 @@ type DecisionRecord struct {
 	Decision  string // what was decided
 	Rationale string // why
 	Context   string // JSON blob of relevant context data
-}
-
-// RetrospectiveRecord holds metadata for a stored retrospective.
-type RetrospectiveRecord struct {
-	ID           int64
-	RunID        string
-	PipelineName string
-	Smoothness   string
-	Status       string // "quantitative", "complete" (has narrative)
-	FilePath     string
-	CreatedAt    time.Time
 }
 
 // Webhook represents a registered webhook endpoint that receives
